Refactor code to allow other programs to access
