+++ bin/common.h
@@ -57,6 +57,26 @@ struct verify_data {
     const char *trusted_host;
 };
 
+struct conn_settings {
+    unsigned mutual_auth:1;
+    unsigned self_service_blinding:1;
+    unsigned only_negotiate:1;
+    unsigned prefer_throughput:1;
+    unsigned prefer_low_latency:1;
+    unsigned enable_mfl:1;
+    unsigned session_ticket:1;
+    unsigned session_cache:1;
+    unsigned insecure:1;
+    unsigned use_corked_io:1;
+    unsigned https_server:1;
+    uint32_t https_bench;
+    int max_conns;
+    const char *ca_dir;
+    const char *ca_file;
+    char *psk_optarg_list[S2N_MAX_PSK_LIST_LENGTH];
+    size_t psk_list_len;
+};
+
 void print_s2n_error(const char *app_error);
 int echo(struct s2n_connection *conn, int sockfd, bool *stop_echo);
 int wait_for_event(int fd, s2n_blocked_status blocked);
@@ -75,3 +95,5 @@ char *load_file_to_cstring(const char *path);
 int s2n_str_hex_to_bytes(const unsigned char *hex, uint8_t *out_bytes, uint32_t max_out_bytes_len);
 int s2n_setup_external_psk_list(struct s2n_connection *conn, char *psk_optarg_list[S2N_MAX_PSK_LIST_LENGTH], size_t psk_list_len);
 uint8_t unsafe_verify_host(const char *host_name, size_t host_name_len, void *data);
+int setup_s2n_connection(struct s2n_connection *conn, int fd, struct s2n_config *config, struct conn_settings settings);
+int s2n_config_settings(int max_early_data, struct s2n_config *config, struct conn_settings conn_settings, const char *cipher_prefs, const char *session_ticket_key_file_path);
